Move impl Arbitrary for Parameter to random/parameter.rs
